requirements.txt
ipykernel==6.29.5
ipywidgets==8.1.3
boto3==1.35.1
sagemaker==2.229.0
huggingface_hub==0.24.6
pydub==0.25.1
aws-cdk-lib==2.142.1
constructs>=10.0.0,<11.0.0
sagemaker==2.229.0